EdgeSolution/modules/image-classifier-service/requirements.txt
pillow
numpy
flask
tensorflow